Updated ManytoMany

+++ Python/ManytoMany/ManytoMany/apps/manytomany/views.py
@@ -4,8 +4,7 @@ from django.contrib import messages
 
 # Create your views here.
 def index(request):
-    context ={"user": User.objects.all()}
-    return render(request, 'manytomany/index.html', context)
+    return render(request, 'manytomany/index.html')
 
 def register(request):
     if request.method == "GET":
@@ -22,12 +21,14 @@ def show(request, id):
     return render(request, 'manytomany/interests.html', context)
 
 def sameinterest(request, id):
+    if request.method == "GET":
+        return redirect('/')
     user = Interest.objects.filter(id = id)
     request.session['interest_name'] = user[0].name
     return redirect('/sameinterest')
 
 def showme(request):
-    context ={"user": User.objects.filter(interests__name= request.session['interest_name'])}
+    context ={"user": User.objects.filter(interests__name = request.session['interest_name'])}
     return render(request, 'manytomany/users.html', context)
 
 def viewpeople(request):
@@ -35,5 +36,7 @@ def viewpeople(request):
     return render(request, 'manytomany/users.html', context)
 
 def delete(request, id):
+    if request.method == "GET":
+        return redirect('/')
     User.objects.filter(id=id).delete()
     return redirect('/viewpeople')
